add booking deletion and booking completed options

+++ HotelMIS/src/main/java/com/WHotels/HotelMIS/model/resort/Booking.java
@@ -28,7 +28,7 @@ public class Booking {
     private Date checkOut;
 
     @Column
-    private BigDecimal total;
+    private Long total;
 
     @ManyToOne
     @JoinColumn(name = "customer_id", referencedColumnName = "customer_id")
